Fix dtype for homogeneity score tests

+++ python/cuml/test/test_metrics.py
@@ -158,8 +158,8 @@ def test_rand_index_score(name, nrows):
 @pytest.mark.parametrize('use_handle', [True, False])
 def test_homogeneity_score(use_handle):
     def score_labeling(ground_truth, predictions):
-        a = np.array(ground_truth, dtype=np.int)
-        b = np.array(predictions, dtype=np.int)
+        a = np.array(ground_truth, dtype=np.int32)
+        b = np.array(predictions, dtype=np.int32)
 
         a_dev = cuda.to_device(a)
         b_dev = cuda.to_device(b)
@@ -206,5 +206,7 @@ def test_homogeneity_score_big_array(use_handle):
 
         np.testing.assert_almost_equal(score, ref, decimal=7)
 
-    assert_ours_equal_sklearn(lambda rng: rng.randint(0, 1000, int(10e4)))
-    assert_ours_equal_sklearn(lambda rng: rng.randint(-1000, 1000, int(10e4)))
+    assert_ours_equal_sklearn(lambda rng: rng.randint(0, 1000, int(10e4),
+                                                      dtype=np.int32))
+    assert_ours_equal_sklearn(lambda rng: rng.randint(-1000, 1000, int(10e4),
+                                                      dtype=np.int32))
